scripts/requirements.txt
python-dotenv
requests
lxml
types-requests
